Change xeng_test to use demux=2
This reflects the xeng demux mode used by the uncommitted xeng_core.mdl file
that was left in fpgadev.njit.edu:/home/observer/projects/eovsa-corr/src.

There seems to be a bug in the casper libraries that requires going into the
xeng block and disabling the link on the xeng_descramble_4ant block so it can
update its parameters (just the annotation string?) when changing the demux
mode.