Updated *_bb.mdl files (about to be deleted)
This commit contains the updated fft_wideband_real_core_bb.mdl and
xeng_core_bb.mdl files from the "recent" (not so recent really, but recently
commited) updates to the corresponding "real" models.  These *_bb.mdl files are
not really necessary, so they are going to be deleted in the next commit, but I
wanted to save them for posterity (just in case!) before doing that.