Add x_op_select_test model and script
This x_op_select_test.mdl file contains the old x_op_select implementation and
the new x_op_select implementation.  The run_x_op_select_test.m script can be
used to setup the input test data and capture the output results.  This can be
used to demonstrate the problem with the old implementation as well as verify
the correctness of the new implementation.