Change x_op_select block for more convenience
Using a shared BRAM in the x_op_select block makes the programming rather
awkward since the correlator model has four such blocks.  All the shared BRAMs
would need to be programmed the same.  This is repetitive and error prone.
Also, shared BRAM do not have init values.

Reverting to a dual port ram that is programmed via a software register bit
banging one port of the BRAM allows one software register to write to all four
BRAMs.  It also allows for an initial value so that the BRAM will come up with
a default output sequence.

Adding a software register to the output of the "programming" port of the BRAM
will provide for software read-back of the BRAM contents.  Just set the address
bits to the desired location, the data bits to anything and the write enable
bit to 0 (obviously but also importantly!).

This model gives the expected results with the run_x_op_select_test.m script.