Optimize part of power_td blocks with DSP48 slices
The power_td blocks used a multiply/add tree to compute the power of four
parallel input samples.  This implementation was not so efficient so it was
replaced with a stack of cascaded DSP48 slices.  There is more room for
optimization here (e.g. by including the accumulation(s) in the stack(s) of
cascaded DSP48 slices), but the current (interim) implementation was done to
avoid having to change any control logic.

The overall latency of the dsp48_sq_add block is 2 cycles longer than the
original multipler/adder implementation.  To accomodate this change in latency,
the sync_in signal has been delayed by a corresponding 2 cycles.