Change BRAM latency in polarimetry block
This is a text editor change to the model file to change the latency of the
"gd_ram" BRAM from 2 to 3.  It also adjusts the latencies of the accomanying
signal paths to keep everything properly synchronized.  The delay changes were
determined by studying the "eval" block and determining where additional
latency was needed to compensate for the increased BRAM delay, then the
requiste chagnes were made in a text editor to make it easy to see where the
changes were made (e.g. by using "git diff").