Set packetizer_x/data_buffer to embedded_reg=on
This sets the "Use embedded registers" option to "on" for the
packetizer_x/data_buffer FIFO.  This should improve clock-to-output times for
that FIFO.  It adds an extra cycle of latency to the FIFO, so the 1-cycle delay
block on the output of the FIFO has been removed.  This delay was not getting
absorbed into the FIFO's embedded registers by the tools, so it wasn't serving
any real purpose.  This change effectively explicitly absorbs that delay into
the FIFO's embedded registers.