Add eovsa_corr_200mhz_2014_Oct_02_1437.bof.gz
This is a new build of the four antenna (8 input) correlator.
It was built for a 200 MHz FPGA clock.

It was built using mlib_devel 150fdcb and includes a newly built xeng.ngc
netlist.

This requires pblock contraints in the UCF file to meet timing.  As of
mlib_devel 150fdcb, these still had to be added manually to the generated UCF
file.