Add x_acc_en control bit
The x_acc_en control bit is the least significant bit of the swreg_ctrl
register.  Until it is set to a 1, data and data_valid samples will be
prevented from reaching the logici downstream from the xeng blocks.  Once that
bit is set to 1, the data and data_valid signals will flow to the downstream
logic.  This is necessary because the downstream logic counts valid samples
without regard for the sync signal, so it is important to prevent that logic
from seeing data_valid=1 prior to synchronizing the ret of the system,
otherwise the valid-counting logic will be out of sync.