Add eovsa_corr_200mhz_2014_Oct_05_1557.bof.gz
This is a build of the four antenna (8 input) correlator.  It was built for a
200 MHz FPGA clock.

It was built using mlib_devel 8005dfb and includes an enable bit (the LSb of
the `sw_reg/ctrl` register).  The enable bit controls the logic downstream from
the `xeng` block.  It is 0 at startup so the controlled logic is disabled until
the user writes a 1 to this bit, which should occur *after* all (both) the
ROACH2s are synchronized with 1 PPS.

In order to meet timing, the tools require some guidance as to where to place
various parts of the design.  This guidance is provided in the form of
`AREA_GROUP` constraints (aka PBlocks), that are stored in the
`eovsa_corr/ucf/pblocks.ucf` file.  As of mlib_devel 8005dfb, all
`<design_name>/ucf/*.ucf` files are automatically included in the constraints
file generated by the CASPER toolflow.