Make 10 GbE reset be level sensitive
This allows the 10 GbE cores (and some related logic) to be held in reset by
writing a '1' to the sw_reg/rst register.  The old behavior was edge sensitive
any only held the 10 GbE reset line high for one clock cycle after the
'0'-to-'1' transition of that bit.  The new behavior can be used to ensure that
no 10 GbE data are transmitted until desired.