Add eovsa_corr_200mhz_2014_Oct_23_1943.bof.gz
This is a build of the four antenna (8 input) correlator.  It was built for a
200 MHz FPGA clock.

It was built using mlib_devel 8005dfb and removes the ineffective use of the
enable bit (the LSb of the `sw_reg/ctrl` register) added in commit abff99e.
It also changes the behavior of the `swreg_rst[0]` bit (aka `rst_ten_gbe`) to
be level sensitive instead of rising edge sensitive.  This allows the 10 GbE
cores to be held in reset for as long as desired under software control.

In order to meet timing, the tools require some guidance as to where to place
various parts of the design.  This guidance is provided in the form of
`AREA_GROUP` constraints (aka PBlocks), that are stored in the
`eovsa_corr/ucf/pblocks.ucf` file.  As of mlib_devel 8005dfb, all
`<design_name>/ucf/*.ucf` files are automatically included in the constraints
file generated by the CASPER toolflow.