Prevent TX oflows when 10 GbE cores are in reset
Apparently the 10 GbE cores will allow the user to shove data into their FIFOs
even when the core's reset input is held high.  This change gates the data
valid inputs to the 10 GbE cores so that it will not be allowed to go high
while the reset line is high.