Add "packet gates" on 10 GbE core inputs
The "packet gates" ensure that incomplete packets are NOT sent out the 10 GbE
interfaces when they are taken out of reset.  When the 10 GbE core comes out of
reset, the remaining part of the packet that is "in progress" (or the next
complete packet if no packet is "in progress") will not be sent out.