Tighten up the `qdr_pblock` area group constraint
This reduces the width of the `qdr_pblock` Pblock so that it does not span any
IOB columns nor any MMCM columns.  The reasoning behind this is that routing
around those large components is presumed to be limited and therefore routing
around them is presumed lead to sub-optimal timing, but that is just a
hypothesis at this point.