Add eovsa_corr_200mhz_2014_Oct_26_1636.bof.gz
This is a build of the four antenna (8 input) correlator.  It was built for a
200 MHz FPGA clock.

It was built using mlib_devel 8005dfb and adds logic to ensure that only
complete packets are sent out the 10 GbE interfaces once they are taken out of
reset.  Without this logic, cores taken out of reset mid-packet would send out
a partial packet.  This would especially confuse the packet recive logic that
counted valid samples without regard for packet completeness (which is by
nature a non-robust design, but there is no need to almost certianly defeat
it).