Add explicit 10 GbE enable bit (swreg_ctrl[0])
The `swreg_ctrol[0]` (lsb) bit is ANDed with `data_valid` in the `pkt_gate`
blocks to provide software control of when valid data words are sent to the
10 GbE cores.  This control bit is ceptured every `eof` pulse so that it will
not take effect until the first valid sample of the packet following the `eof`
sample.  This ensures that transmission will only be enabled/disabled on packet
boundaries.