Add eovsa_corr_200mhz_2014_Oct_29_0021.bof.gz
This is a build of the four antenna (8 input) correlator.  It was built for a
200 MHz FPGA clock.

It was built using mlib_devel 8005dfb and adds a 10 GbE enable bit that must be
explicitly set to 1 via software before packets will be sent.  This enable bit
(`swreg_ctrl[0]`) is registered locally at each 10 GbE core by the core's `eof`
signal to ensure that changes it the bit's state only take effect on packet
boundaries.