Add eovsa_fx_snoop.mdl
The eovsa_fx_snoop model is designed to snoop on the FX dataflow between
ROACH2s.  It is a test bed for features that could be included in the full
eovsa_corr design (if it doesn't blow up the timing).  The eovsa_fx_snoop model
transmits no packets, but it will receive packets and facilities exists to
capture packets.  It should be run on one (or more) ROACH2(s) of the eovsa
correlator with at least one ROACH2 running the eovsa_corr model.  The current
eovsa_corr model only runs on two ROACH2s so this design should only be run on
one of them at most.