Update packet_buffer readout logic
Implement new packet buffer receive logic that synchronizes all the X engines
in a ROACH with each other and allows for synchronizing integrations across
multiple ROACHes by ensuring that they all start integrating on the same
mcount.

The current design/implementation is complicated because data are not sent
continuously.  There are gaps in the data flow during the blanking interval
when the LO changes.  It would be easier (from a data flow point of view) if
there were no gaps and zeros were sent during the blanking interval.  I was
going to make this change, but then I realized that the "corner turn" (matrix
transpose) that happens after the blanking logic relies on these gaps for its
control logic.  I didn't want to change the corner turner, so I abandoned the
idea of sending zeros and instead decided to deal with the gaps.

This commit introduces a new "packet buffer read out" state machine to handle
these gaps that is cleaner and better documented than the old one.
Additionally, and perhaps most significantly, things are setup so that one
packet buffer's readout logic will drive the readout of all four packet
buffers.  Using one of the packet buffers as a master for all will ensure that
the X engines are all driven in unison.  A new `swreg_mcount_start` register
needs to be programmed at start-up with an mcount that will be in the near
future.  This value determines the mcount at which data will start flowing to
the X engines.  This is used to synchronize integrations across multiple
boards.  This is a 32-bit register that specifies the top 32 bits of the
36-bit mcount (the lower 4 bits are always treated as 0).

The upper 32 bits of the 36-bit mcount most recently received from the two 10
GbE interfaces can be read from the new `rx_mcount_fx0` and `rx_mcount_fx1`
registers.  Both of these registers should increment steadily and in near
unison.  The `rx_mcount_fx0` register contains the value that is compared
against `swreg_mcount_start` to determine when to start reading out the packet
buffers.

Reading the new `rx_mcounts` register will return the lower 16 bits of the most
recently received mcounts from both 10 GbE interfaces.  The lower 16 bits of
the FX0 mcount will be in the upper 16 bits; the lower 16 bits of the FX1
mcount will be in the lower 16 bits.  The two mcounts should be very close to
each other.  The `rx_mcounts` register captures the lower portion of both
mcount values simultaneously, so this register can be used to validate that
both 10 GbE cores are receiving synchronized data.

Also add pktbuf_state_diagram.dot which contains the Graphviz "dot" source code
for the packet buffer state diagram.  The pktbuf_state_diagram.png file
contains an image of the state machine diagram generated with:

    dot -Tpng -o pktbuf_state_diagram.png pktbuf_state_diagram.dot

This PNG file is shown in one of the new packet buffer state machine blocks.