Use P packetizer acc_num for both P and X packets
The P and X packetizers used to keep indpendent but supposeduly (I have my
doubts) equivalent accumulation counts.  This "worked" so long as the P and X
parts of the design started their accumulation counters at the same time.  Now
that the FX packet receive logic does not start correlations and therefore
accumulations until a specified mcount is reached, the P and X parts of the
design no longer start their accumulation counters at the same time.  This
leads to different acc_num values in P and X packets that correspond in time.

This commit uses the P packetizer acc_num value for both P and X packets.  The
X packetizer capture the P packetizer acc_num value whenever it is sending out
packet 0 of an accumulation and uses that captured acc_num value for all
packets of that X accumulation.

As an editorial comment, the entire design could use a more traceable sync and
timestamp architecture.  The current implementation is disjointed with "sync"
signals often being discarded only to be regenerated downstream by counting
"valid" samples.