Add `reorder_test.mdl`
This model file is used to simulate the F data corner turn and packetization.

The `pkt1_scope` block shows a mismatch between the packet header and the
packet contents.  Here is a table with three columns: packet contents, current
(bad) header, expected (good) header.  The headers are shown as triplets of
numbers (timestamp, channel, antenna).

      Data    Current Header   Expected Header
    ------- ----------------- -----------------
    a0/c000 (t=0, c=000, a=0) (t=0, c=000, a=0)
    a1/c000 (t=0, c=000, a=0) (t=0, c=000, a=1) ***
    a0/c800 (t=0, c=000, a=1) (t=0, c=800, a=0)
    a1/c800 (t=0, c=800, a=0) (t=0, c=800, a=1)
    a0/c200 (t=0, c=800, a=1) (t=0, c=200, a=0)
    a1/c200 (t=0, c=200, a=0) (t=0, c=200, a=1)
    a0/cA00 (t=0, c=200, a=1) (t=0, c=A00, a=0)
    a1/cA00 (t=0, c=A00, a=0) (t=0, c=A00, a=1)

    a0/c001 (t=0, c=A00, a=1) (t=0, c=001, a=0)
    a1/c001 (t=0, c=001, a=0) (t=0, c=001, a=1)
    ...

The problem is that the header for the very first packet gets duplicated for
the second packet (marked with `***` above).  From then on all headers are off
by one packet.

Another problem exists in the packetizer blocks.  The counter used to generate
the packet headers is reset by either the incoming sync signal or the
asynchronous (relative to the data) `mrst` signal from a software register.

Another problem in this design is the way that sync is propagated through the
`ct` block.  The `ct` block's `sync_out` signal is a one-shot pulse generated
by the very first rising edge of the `valid0` output.  This might happen well
before the multiple FPGAs in the system have been synchronized.

One solution to both problems is to have the `ct` block generate a `sync_out`
pulse for the first valid cycle of each "slot's" output (there are 50 "slots"
per second, each corresponding to an LO setting).  Then the packetizer blocks
could use the `mrst` signal to arm the header counter to reset on the next
sync pulse from the `ct` block.  To synchronize across boards, the sync cannot
happen for every 20 ms slot; it should happen for only the first slot of a
second.  This requires additional info from `new_acc` block upstream from the
`ct` block.

Future versions of the `reorder_test.mdl` file will incorporate fixes for these
problems.