Fix F header/data mismatch in reorder_test.mdl
This fixes the off-by-one-packet mismatch between the F packet header and the F
packet payload data.  The fix, in the `packetiser0` and `packtizer1` blocks, is
to use the internal `header_valid` signal to advance the `clk_cnt` packet
counter rather than the internal `eof_out` signal.  This ensures that the
proper header value is ready in time for the next packet.

Still to come is the fix for the reset of the `clk_cnt` counter.