Fix packet counter reset logic in reorder_test.mdl
This implements the following changes to fix the reset of the `clk_cnt`
counters inside the `packetiser0` and `packetiser1` blocks of the
`reorder_test.mdl` model.

Passes the new `slot_0` signal output from `new_acc` into the `ct` block.  Like
the rest of the `new_acc` signals, the `slot_0` signal in the `eovsa_corr`
model is derived from the global `sync` signal, which is a one-time armed pulse
from 1 PPS.

The `ct` block now generates its `sync_out` signal one cycle before the first
rising edge of `valid0` after a falling edge of `slot_0`.  This ensures that
the `sync_out` pulse happens at the beginning of every slot 0 (i.e. the first
slot of each second).

The packtizers now pass this sync signal through an `armed_trigger` block whose
output is used to reset the `clk_cnt` counter.  The `armed_trigger` block is
armed via the `mrst` signal, which is generated from the rising edge of
`sw_reg/reset[16]`.  This ensures that the packetizers' `clk_cnt` counters will
always be "in phase" with the data and the slots, even when `mrst` is used to
reset the counters.

One final fix is to use the packetizers' internal `header_valid` signal to
increment the `clk_cnt` counter rather than the internal `eof_out` signal.