Sync `f_x` subsystem using `sync_polarimetry_out`
The `f_x` subsystem had been synchronized using the `sync` goto tag, but this
is directly from the `timing` subsystem despite that fact that the `f_x`
subsystem's input data is from the output of the `polarimetry` subsystem.

This commit changes the sync input to the `f_x` subsystem to be the sync output
of the `polarimetry` subsystem.  This ensures proper alignment between sync and
data going into the `f_x` subsystem block.