Add eovsa_corr_2015_May_19_0046.bof.gz
This includes the change to the `sync_in` input of the `f_x` block to use the
`sync_out` of the `polarimetry` block to ensure proper sync/data alignment
going into the `f_x` block.  This is intended/expected (really!) to fix the
"channel offset" problem seen in the F packets (and downstream in the X engine
output).

This was built using mlib_devel 374f819.