Add files for QDR vacc
This is not the standard casper library block. It uses
a verilog state machine to control the QDR interface, and
operates based on an assumed input burst pattern.
I.e., the block assumes that data will be written into the vacc
in a burst of N words, and that between bursts there are a continuous
block of M clocks during which the last accumulation in the vacc
can be read out.