Nominalluy complete 16a design
This design features all the major parts required for operation
with 16 antennas.

It features 4 X-engine cores, with 256 sample accumulators, which
are muxed into a single QDR vacc. Output of the vacc is in bursts
of 16*17/2 * 4 (i.e. all) baselines, with one such burst in each
output UDP packet.

This design has not been compiled.