Make X-engine outputs continuous bursts
This is implemented by re-buffering the xengine
outputs, which are valied every N clocks. It should
be possible to do this in the descramble, but it's
not entirely trivial. The HDL version of the xengine
block (as used in AMI, not necessarily as stored in the
library, can do this)