Get X-header global_acc field from P-Packets
This is what the 4 antenna design did. As the X-engine
delivers a new accumulation to the X packetiser, the P-packet
global_acc_count is grabbed and used for that acc's X-outputs.

Since an X-accumulation is only ready to output after all it's data has
been processed, it doesn't seem obvious that the P-packetizer won't
have moved on to the next accumulation by the time the X-packetizer
grabs the global_acc_count. Beware.