Fix channel order bug. Don't EQ P/P^2 data
This commit implements two changes.

1. Don't EQ the P/P^2 data stream. The 18-bit stream used for
feeding the power/power-squared computation block is now driven
directly from the FFT outputs. It no longer comes from the EQ'd
stream that feeds the X-engine 4-bit quantizers. This is because
the EQ necessarily has to scale so that high RFI-bins saturate
in order to keep the non-RFI bins in a reasonable power level.

2. A block within the F->X corner-turner had not been correctly
updated for the 8-ROACH system, and as a result was misordering
frequency channels. This commit fixes this.